# docs/swagger.yaml
basePath: /
definitions:
  domain.CreateProductParams:
    properties:
      category:
        example: Инструменты
        type: string
      count:
        example: 50
        type: integer
      manufacturer:
        example: Зубр
        type: string
      price:
        example: 1299
        type: integer
      product_name:
        example: Молоток
        type: string
    type: object
  domain.PatchProductParams:
    properties:
      category:
        example: Инструменты
        type: string
      count:
        example: 30
        type: integer
      manufacturer:
        example: Зубр
        type: string
      price:
        example: 1499
        type: integer
      product_name:
        example: Молоток с резиновой ручкой
        type: string
    type: object
  domain.Product:
    properties:
      category:
        example: Инструменты
        type: string
      count:
        example: 50
        type: integer
      id:
        example: 42
        type: integer
      manufacturer:
        example: Зубр
        type: string
      price:
        example: 1299
        type: integer
      product_name:
        example: Молоток
        type: string
    type: object
  handler.CreateProductResponse:
    properties:
      id:
        type: integer
    type: object
  handler.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  handler.refreshRequest:
    properties:
      refresh:
        type: string
    required:
    - refresh
    type: object
  service.TokenPair:
    properties:
      access:
        type: string
      refresh:
        type: string
    type: object
info:
  contact: {}
  description: API для управления складом
  title: Warehouse Controller API
  version: "1.0"
paths:
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Принимает действующий refresh-токен и возвращает новую пару
      parameters:
      - description: Refresh-токен
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.refreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.TokenPair'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Обновить пару токенов
      tags:
      - auth
  /auth/token:
    post:
      description: Создаёт новую пару access/refresh токенов без проверки учётных
        данных
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.TokenPair'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Выдать пару токенов
      tags:
      - auth
  /products:
    get:
      parameters:
      - description: Лимит результатов
        in: query
        name: limit
        type: integer
      - description: Смещение
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Product'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Список продуктов
      tags:
      - products
    post:
      consumes:
      - application/json
      parameters:
      - description: Параметры нового продукта
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.CreateProductParams'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.CreateProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать продукт
      tags:
      - products
  /products/{id}:
    delete:
      parameters:
      - description: ID продукта
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить продукт (soft delete)
      tags:
      - products
    get:
      parameters:
      - description: ID продукта
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить продукт по ID
      tags:
      - products
    patch:
      consumes:
      - application/json
      parameters:
      - description: ID продукта
        in: path
        name: id
        required: true
        type: integer
      - description: Поля для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.PatchProductParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Частичное обновление продукта
      tags:
      - products
  /products/{id}/restore:
    put:
      parameters:
      - description: ID продукта
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Восстановить ранее удалённый продукт
      tags:
      - products
  /products/search:
    get:
      description: Фильтрация продуктов по имени, производителю, категории и диапазону
        цены
      parameters:
      - description: Подстрока в названии
        in: query
        name: product_name
        type: string
      - description: Производитель
        in: query
        name: manufacturer
        type: string
      - description: Категория
        in: query
        name: category
        type: string
      - description: Минимальная цена
        in: query
        name: min_price
        type: integer
      - description: Максимальная цена
        in: query
        name: max_price
        type: integer
      - description: Лимит результатов
        in: query
        name: limit
        type: integer
      - description: Смещение
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Product'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Поиск продуктов
      tags:
      - products
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"